integration

+++ microservice/api_produto_microservice/produto_microservice/src/main/java/com/produto/service/produto_microservice/dto/ProdutoDto.java
@@ -1,4 +1,4 @@
 package com.produto.service.produto_microservice.dto;
 
-public record ProdutoDto(Long id, Long id_armazen, String nome, Long quantidade, Float price) {
+public record ProdutoDto(Long id, Long id_armazen, String nome, Integer quantidade, Float price) {
 }
